requirements.txt
speechrecognition
pyaudio
pyttsx3
pywhatkit
wikipedia
beautifulsoup4
lxml
requests
python-dotenv
pyautogui
pygetwindow
